present: Introduce internal flip mode API
The current mode of operation only allows flips per screen.
In order to allow other more advanced modes (such as flips per window),
introduce an internal API for adding and maintaining those modes easily.

A mode "mde" is selected by calling the approriate present_mdemode_init
function on a particular ScreenPtr. Internally then several function
hooks need to be set. When possible common code between modes is in
present.c and diverging functionality shall go in present_mdemode.c.

This patch only adds all the API hooks and rewires the currently only
supported screen flipping as the first available mode in the new API.
There is otherwise no change in functionality and the current workings
are meant to stay completely the same.

The scrmode is also used, when flipping is not supported by the hardware.

Signed-off-by: Roman Gilg <subdiff@gmail.com>