xwayland: Preliminary support for present's new window flip mode
Introduce support for present's window flip mode. The support is not yet
complete, but works reasonable well for the most important use case, what
is fullscreen applications

We take a present flip and if the xwl_window->window has the same dimensions
as the presenting window, the flip is represented by a wl_buffer and
attached to the main wl_surface of the xwl_window.

After commit we are listening for the sync callback in order to tell present,
that the pixmap flip is not longer pending, for the frame callback in order
to update the msc counter and for the buffer release callback in order to tell
present that the pixmap is idle again.

In case the compositor is not sending any more frame callbacks, we use a slow
timer to continue the msc counter.

The following functionality is missing from this implementation:
* per window flips for child windows with smaller size than the xwl_window,
* queuing events to MSC times,
* reporting UST values to present.

To make use of this functionality Xwayland must run rootless and with
Glamor/GBM.

Signed-off-by: Roman Gilg <subdiff@gmail.com>