xwayland: Cut off upper 32bit of queued vblank delay
This ensures the same behavior as in Present's fake counter and xfree86.

At the moment clients might do put vblanks too far into the future, because
the fake vblank code in Present and the xfree86 driver tolerate cut off upper
32bit due to an 64 to 32bit conversion. Do this therefore here as well to not
suddenly regress on Xwayland only.

The sample client, that triggers this behavior, is the Steam client.

Signed-off-by: Roman Gilg <subdiff@gmail.com>