xwayland: Guard against very late vblanks
Do not allow queuing events too far into the the future. The result can be a
presentation freeze until the msc time is reached.

At the moment clients might do this per accident, because the fake vblank code
in Present and the xfree86 driver tolerate high 64bit msc values without
freeze due to an erroneous 64 to 32bit conversion.

The sample client, that triggers this behavior, is the Steam client.

Signed-off-by: Roman Gilg <subdiff@gmail.com>